Extract History component
